requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
crontab==1.0.4
cryptography==44.0.0
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-ipware==7.0.1
django-redis==5.4.0
django-rest-authemail==2.1.7
django-rq==3.0.1
djangorestframework==3.15.2
freezegun==1.5.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
paramiko==3.5.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==3.0.0
redis==6.2.0
requests==2.32.3
rq==2.3.3
rq-scheduler==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0